behaviors.JSON -> Local; fixed COPY; removed filter api tutorial
