+++ src/Hypergrid/events.js
@@ -499,7 +499,7 @@ var mixin = {
 
         //Register a listener for the copy event so we can copy our selected region to the pastebuffer if conditions are right.
         document.body.addEventListener('copy', function(evt) {
-            self.checkClipboardCopy(evt);
+            grid.checkClipboardCopy(evt);
         });
     },
