+++ src/Hypergrid/index.js
@@ -17,7 +17,7 @@ var Renderer = require('../renderer');
 var SelectionModel = require('../lib/SelectionModel');
 var Localization = require('../lib/Localization');
 var Behavior = require('../behaviors/Behavior');
-var behaviorJSON = require('../behaviors/JSON');
+var behaviorJSON = require('../behaviors/Local');
 var cellRenderers = require('../cellRenderers');
 var cellEditors = require('../cellEditors');
 var modules = require('./modules');
@@ -61,7 +61,7 @@ var EDGE_STYLES = ['top', 'bottom', 'left', 'right'],
  *
  * @param {string|Element} [container] - CSS selector or Element. If omitted (and `options.container` also omitted), Hypergrid first looks for an _empty_ element with an ID of `hypergrid`. If not found, it will create a new element. In either case, the container element has the class name `hypergrid-container` added to its class name list. Finally, if the there is more than one such element with that class name, the element's ID attribute is set to `hypergrid` + _n_ where n is an ordinal one less than the number of such elements.
  * @param {object} [options] - If `options.data` provided, passed to {@link Hypergrid#setData setData}; else if `options.Behavior` provided, passed to {@link Hypergrid#setBehavior setBehavior}.
- * @param {function} [options.Behavior=behaviors.JSON] - _Per {@link Behavior#setData}._
+ * @param {function} [options.Behavior=Local] - _Per {@link Behavior#setData}._
  * @param {dataModelAPI} [options.dataModel] - _Passed to behavior {@link Behavior constructor}._
  * @param {function} [options.DataModel=require('datasaur-local')] - _Passed to behavior {@link Behavior constructor}._
  * @param {function|object[]} [options.data] - _Passed to behavior {@link Behavior constructor}._
@@ -609,7 +609,7 @@ var Hypergrid = Base.extend('Hypergrid', {
      * @param {object} [options]
      * @param {string[]} [options.blacklist] - List of grid properties to exclude. Pertains to grid own properties only.
      * @param {boolean} [options.compact] - Run garbage collection first. The only property this current affects is `properties.calculators` (removes unused calculators).
-     * @param {number|string} [options.space='\t'] - For no space, give `0`. (See {@link https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/JSON/stringify|JSON.stringify}'s `space` param other options.)
+     * @param {number|string} [options.space='\t'] - For no space, give `0`. (See {@link https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Local/stringify|JSON.stringify}'s `space` param other options.)
      * @param {function} [options.headerify] - If your headers were generated by a function (taking column name as a parameter), give a reference to that function here to avoid persisting headers that match the generated string.
      * @memberOf Hypergrid#
      */
@@ -787,7 +787,7 @@ var Hypergrid = Base.extend('Hypergrid', {
      * * Hypergrid constructor
      * * `setData` when not called explicitly before then
      * @param {object} [options] - _Per {@link Behavior#setData}._
-     * @param {Behavior} [options.Behavior=behaviors.JSON] - The behavior (model) can be either a constructor or an instance.
+     * @param {Behavior} [options.Behavior=Local] - The behavior (model) can be either a constructor or an instance.
      * @param {dataModelAPI} [options.dataModel] - A fully instantiated data model object.
      * @param {function} [options.DataModel=require('datasaur-local')] - Data model will be instantiated from this constructor unless `options.dataModel` was given.
      * @param {dataModelAPI} [options.metadata] - Value to be passed to setMetadataStore if the data model has changed.
@@ -803,6 +803,48 @@ var Hypergrid = Base.extend('Hypergrid', {
     },
 
     /**
+     * Number of _visible_ columns.
+     * @memberOf Hypergrid#
+     * @returns {number} The number of columns.
+     */
+    getColumnCount: function() {
+        return this.behavior.getActiveColumnCount();
+    },
+
+    /**
+     * @memberOf Hypergrid#
+     * @returns {number} The number of rows.
+     */
+    getRowCount: function() {
+        return this.behavior.getRowCount();
+    },
+
+    getRow: function(y) {
+        return this.behavior.getRow(y);
+    },
+
+    /**
+     * @memberOf Hypergrid#
+     * @summary Get data value at given cell.
+     * @param {number} x - The horizontal coordinate.
+     * @param {number} y - The vertical coordinate.
+     */
+    getValue: function(x, y) {
+        return this.behavior.getValue.apply(this.behavior, arguments); // must use .apply (see this.behavior.getValue)
+    },
+
+    /**
+     * @memberOf Hypergrid#
+     * @summary Set a data value of a given cell.
+     * @param {number} x - The horizontal coordinate.
+     * @param {number} y - The vertical coordinate.
+     * @param {*} value - New cell value.
+     */
+    setValue: function(x, y, value) {
+        this.behavior.setValue.apply(this.behavior, arguments); // must use .apply (see this.behavior.setValue)
+    },
+
+    /**
      * @memberOf Hypergrid#
      * @summary Set the underlying datasource.
      * @desc This can be done dynamically.
@@ -810,7 +852,7 @@ var Hypergrid = Base.extend('Hypergrid', {
      * * An array of congruent raw data objects.
      * * A function returning same.
      * @param {object} [options] - _(See also {@link Behavior#setData} for additional options.)_
-     * @param {Behavior} [options.Behavior=behaviors.JSON] - The behavior (model) can be either a constructor or an instance.
+     * @param {Behavior} [options.Behavior=Local] - The behavior (model) can be either a constructor or an instance.
      * @param {dataModelAPI} [options.dataModel] - _Passed to behavior {@link Behavior constructor} (when `options.Behavior` given)._
      * @param {function} [options.DataModel=require('datasaur-local')] - _Passed to behavior {@link Behavior constructor} (when `options.Behavior` given)._
      * @param {dataModelAPI} [options.metadata] - _Passed to behavior {@link Behavior constructor} (when `options.Behavior` given)._
@@ -1339,23 +1381,6 @@ var Hypergrid = Base.extend('Hypergrid', {
     },
 
     /**
-     * Number of _visible_ columns.
-     * @memberOf Hypergrid#
-     * @returns {number} The number of columns.
-     */
-    getColumnCount: function() {
-        return this.behavior.getActiveColumnCount();
-    },
-
-    /**
-     * @memberOf Hypergrid#
-     * @returns {number} The number of rows.
-     */
-    getRowCount: function() {
-        return this.behavior.getRowCount();
-    },
-
-    /**
      * @memberOf Hypergrid#
      * @returns {number} The number of fixed columns.
      */
@@ -1683,9 +1708,6 @@ var Hypergrid = Base.extend('Hypergrid', {
     },
     lookupFeature: function(key) {
         return this.behavior.lookupFeature(key);
-    },
-    getRow: function(y) {
-        return this.deprecated('getRow(y)', 'behavior.dataModel.getRow(y)', '3.0.0', arguments, 'We removed grid.getRow(y) and grid.behavior.getRow(). If you are determined to call getRow, call it on the data model directily. However, calling .getRow(y) is not recommended; always try to use .getValue(x, y) instead. See https://github.com/fin-hypergrid/core/wiki/getRow(y)-and-getData()-(ab)use for more information.');
     },
 
     newPoint: function(x, y) {
